feat(api): initial work on api
