--- api/cbor_handlers.go
@@ -1,20 +1,0 @@
-package api
-
-import (
-	"net/http"
-)
-
-type CBORHandlers struct {
-}
-
-func (h *CBORHandlers) HandleListWalks(w http.ResponseWriter, r *http.Request) {
-
-}
-
-func (h *CBORHandlers) HandleWalk(w http.ResponseWriter, r *http.Request) {
-
-}
-
-func (h *CBORHandlers) HandleResource(w http.ResponseWriter, r *http.Request) {
-
-}
